refactor: migrate frontend tooling and update dependencies
