--- frontend/src/index.js
@@ -1,19 +1,0 @@
-// /frontend/src/index.js
-
-import React from 'react';
-import ReactDOM from 'react-dom/client';
-import './index.css';
-import App from './App';
-import reportWebVitals from './reportWebVitals';
-import { AuthProvider } from './AuthContext'; // Import AuthProvider
-
-const root = ReactDOM.createRoot(document.getElementById('root'));
-root.render(
-  <React.StrictMode>
-    <AuthProvider> {/* Wrap App with AuthProvider */}
-      <App />
-    </AuthProvider>
-  </React.StrictMode>
-);
-
-reportWebVitals();
