+++ frontend/src/pages/HistoryCard.jsx
